refactor: start runtime earlier
