+++ deployer/src/lib.rs
@@ -1,4 +1,4 @@
-use std::{convert::Infallible, env, net::SocketAddr, path::PathBuf};
+use std::{convert::Infallible, net::SocketAddr};
 
 pub use args::Args;
 pub use deployment::{deploy_layer::DeployLayer, runtime_logger::RuntimeLoggerFactory};
@@ -10,7 +10,6 @@ use hyper::{
 };
 pub use persistence::Persistence;
 use proxy::AddressGetter;
-use tokio::select;
 use tracing::{error, info};
 
 mod args;
@@ -45,27 +44,10 @@ pub async fn start(persistence: Persistence, args: Args) {
     );
     let make_service = router.into_make_service();
 
-    let workspace_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .parent()
-        .unwrap()
-        .to_path_buf();
-
-    let runtime_dir = workspace_root.join("target/debug");
-
-    let mut runtime = tokio::process::Command::new(runtime_dir.join("shuttle-runtime"))
-        .args(&["--legacy", "--provisioner-address", "http://localhost:8000"])
-        .current_dir(&runtime_dir)
-        .spawn()
-        .unwrap();
-
-    select! {
-        _ = runtime.wait() => {
-            info!("Legacy runtime stopped.")
-        },
-        _ = axum::Server::bind(&args.api_address).serve(make_service) => {
-            info!("Handlers server stopped serving addr: {}", &args.api_address);
-        },
-    }
+    axum::Server::bind(&args.api_address)
+        .serve(make_service)
+        .await
+        .unwrap_or_else(|_| panic!("Failed to bind to address: {}", args.api_address));
 }
 
 pub async fn start_proxy(
